Implement JWT auth via HttpOnly cookie and refactor note/user DTO mapping

+++ src/main/java/com/example/collabnotes/service/NoteService.java
@@ -1,9 +1,10 @@
 package com.example.collabnotes.service;
 
-import com.example.collabnotes.controller.AuthController;
+import com.example.collabnotes.dto.CreateNoteRequest;
 import com.example.collabnotes.dto.NoteResponse;
 import com.example.collabnotes.entity.Note;
 import com.example.collabnotes.entity.User;
+import com.example.collabnotes.mapper.CreateNoteRequestMapper;
 import com.example.collabnotes.mapper.NoteResponseMapper;
 import com.example.collabnotes.repository.NoteRepository;
 import com.example.collabnotes.repository.UserRepository;
@@ -13,9 +14,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -24,8 +23,12 @@ public class NoteService {
     private final NoteRepository noteRepository;
     private final UserRepository userRepository;
     private final NoteResponseMapper noteResponseMapper;
+    private final CreateNoteRequestMapper createNoteRequestMapper;
 
     public List<NoteResponse> getAllNotes(Long userId) {
+
+        Authentication authentication = SecurityContextHolder.getContext()
+                .getAuthentication();
 
         return (userId != null ? noteRepository.findAllByCreatedBy_Id(userId) : noteRepository.findAll()).stream()
                 .map(noteResponseMapper::map)
@@ -39,23 +42,29 @@ public class NoteService {
         return noteResponseMapper.map(note);
     }
 
-    public NoteResponse createNote(Note note) {
+    public NoteResponse createNote(CreateNoteRequest createNoteRequest) {
         Authentication authentication = SecurityContextHolder.getContext()
                 .getAuthentication();
-        User user = userRepository.findByEmail(authentication.getName())
-                .orElseThrow(() -> new EntityNotFoundException("User with id: " + authentication.getName() + " not found"));
 
-        Note noteToCreate = Note.builder()
-                .title(note.getTitle())
-                .content(note.getContent())
-                .createdBy(user)
-                .build();
+        String email = authentication.getName();
 
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new EntityNotFoundException("User with mail: " + email + " not found"));
+
+        Note noteToCreate = createNoteRequestMapper.map(createNoteRequest);
+        noteToCreate.setCreatedBy(user);
         Note noteCreated = noteRepository.save(noteToCreate);
         return noteResponseMapper.map(noteCreated);
     }
 
     public void deleteNoteById(Long noteId) {
+
+        boolean isExist = noteRepository.existsById(noteId);
+
+        if (!isExist) {
+            throw new EntityNotFoundException("Note with id: " + noteId + " not found");
+        }
+
         noteRepository.deleteById(noteId);
     }
 }
